backend/fastapi-rag/pyproject.toml
[project]
name = "fastapi-rag"
version = "0.1.0"
description = "Miguel's RAG Assistant with ChromaDB"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "openai>=1.3.7",
    "python-dotenv>=1.0.0",
    "chromadb>=0.4.22",
    "aiohttp>=3.12.15",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.24.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Minimal configuration - just specify the app directory
[tool.hatch.build.targets.wheel]
packages = ["app"]

[project.scripts]
fastapi-rag = "app.main:main"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.24.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
